moved CloudWatchWriter to XWriterFactory and XWriter design
